BAEL-3506
