+++ java-math-2/README.md
@@ -1,0 +1,8 @@
+## Java Math
+
+This module contains articles about math in Java.
+
+### Relevant articles:
+
+- [Basic Calculator in Java](https://www.baeldung.com/basic-calculator-in-java)
+- More articles: [[<-- prev]](/../java-math)
