+++ java-math-2/src/main/java/com/baeldung/maths/calculator/basic/BasicCalculatorSwitchCase.java
@@ -1,21 +1,21 @@
-package com.baeldung.maths;
+package com.baeldung.maths.calculator.basic;
 
 import java.util.InputMismatchException;
 import java.util.Scanner;
 
-public class BasicCalculatorIfElse {
-
+public class BasicCalculatorSwitchCase {
     public static void main(String[] args) {
 
-        Scanner scanner = new Scanner(System.in);
-        System.out.println("---------------------------------- \n" +
-                "Welcome to Basic Calculator \n" +
-                "----------------------------------");
+        System.out.println("---------------------------------- \n"
+            + "Welcome to Basic Calculator \n"
+            + "----------------------------------");
         System.out.println("Following operations are supported : \n" +
                 "1. Addition (+) \n" +
                 "2. Subtraction (-) \n" +
                 "3. Multiplication (* OR x) \n" +
                 "4. Division (/) \n");
+
+        Scanner scanner = new Scanner(System.in);
         try {
             System.out.println("Enter an operator: (+ OR - OR * OR /) ");
             char operation = scanner.next().charAt(0);
@@ -34,19 +34,31 @@ public class BasicCalculatorIfElse {
                 System.err.println("Second Number cannot be zero for Division operation.");
             }
 
-            if (operation == '+') {
-                System.out.println(num1 + " + " + num2 + " = " + (num1 + num2));
-            } else if (operation == '-') {
-                System.out.println(num1 + " - " + num2 + " = " + (num1 - num2));
-            } else if (operation == '*' || operation == 'x') {
-                System.out.println(num1 + " x " + num2 + " = " + (num1 * num2));
-            } else if (operation == '/') {
-                System.out.println(num1 + " / " + num2 + " = " + (num1 / num2));
-            } else {
-                System.err.println("Invalid Operator Specified.");
+            switch (operation) {
+                case '+':
+                    System.out.println(num1 + " + " + num2 + " = " + (num1 + num2));
+                    break;
+                case '-':
+                    System.out.println(num1 + " - " + num2 + " = " + (num1 - num2));
+                    break;
+                case '*':
+                    System.out.println(num1 + " x " + num2 + " = " + (num1 * num2));
+                    break;
+                case 'x':
+                    System.out.println(num1 + " x " + num2 + " = " + (num1 * num2));
+                    break;
+                case '/':
+                    System.out.println(num1 + " / " + num2 + " = " + (num1 / num2));
+                    break;
+                default:
+                    System.err.println("Invalid Operator Specified.");
+                    break;
             }
         } catch (InputMismatchException exc) {
             System.err.println(exc.getMessage());
+        } finally {
+            scanner.close();
         }
     }
-} 
+}
+
